requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Image processing and AI
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
scikit-learn==1.3.2

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0

# Raspberry Pi GPIO and sensors
RPi.GPIO==0.7.1
adafruit-circuitpython-dht==3.7.9
w1thermsensor==2.0.0
smbus2==0.4.2

# Database and data handling
sqlite3
pandas==2.1.4
numpy==1.25.2

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Additional utilities
aiohttp==3.9.1
aiofiles==23.2.1
click==8.1.7
colorama==0.4.6
rich==13.7.0
tqdm==4.66.1
schedule==1.2.0
watchdog==3.0.0
python-dateutil==2.8.2
psutil==5.9.6
webdriver-manager==4.0.1
jinja2==3.1.2
httpx==0.25.2
cachetools==5.3.2
ratelimit==2.2.1
typer==0.9.0
nltk==3.8.1
textblob==0.17.1
lxml==4.9.3
colorthief==0.2.1
exifread==3.0.0
orjson==3.9.10
cryptography==41.0.8
distro==1.8.0
environs==10.0.0
pyjwt==2.8.0
email-validator==2.1.0
slowapi==0.1.9
prometheus-client==0.19.0
cerberus==1.3.5
factory-boy==3.3.0
faker==20.1.0
ipdb==0.13.13
ipython==8.17.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
xxhash==3.4.1
dill==0.3.7
joblib==1.3.2